requirements.txt
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
Jinja2==2.11.2
python-dotenv==0.16.0
Requests==2.25.1
WTForms==2.3.3
